migrate types for user controller/signup to types folder
